# openapi.yaml
openapi: 3.1.0
info:
  title: RESTful API Challenge
  description: Base Spring Boot RESTful API with standard architecture patterns
  contact:
    name: Developer
    email: dev@challenge.com
  license:
    name: MIT License
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local Development
tags:
- name: Transactions
  description: CRUD operations for financial transactions
paths:
  /transactions/{id}:
    get:
      tags:
      - Transactions
      summary: Get transaction by id
      operationId: findById
      parameters:
      - name: id
        in: path
        description: Transaction id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    description: Transaction response payload
                    properties:
                      id:
                        type: integer
                        format: int64
                        example: 1
                      amount:
                        type: number
                        format: double
                        example: 150.0
                      type:
                        type: string
                        example: car
                      parent_id:
                        type: integer
                        format: int64
                        example: 0
                  timestamp:
                    type: string
                    format: date-time
        "404":
          description: Transaction not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: false}
                  message: {type: string, example: "Transaction not found: 123"}
                  timestamp: {type: string, example: "2026-04-19T17:56:17"}
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: false}
                  message: {type: string, example: "An unexpected error occurred"}
                  timestamp: {type: string, example: "2026-04-19T17:56:17"}
    put:
      tags:
      - Transactions
      summary: Create or update a transaction
      operationId: putTransaction
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Payload to create or update a transaction
              properties:
                amount:
                  type: number
                  format: double
                  example: 150.0
                type:
                  type: string
                  example: car
                parent_id:
                  type: integer
                  format: int64
                  example: 0
              required:
              - amount
              - type
        required: true
      responses:
        "200":
          description: OK (Transaction updated)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        "201":
          description: Created (Transaction created)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        "400":
          description: Bad request (Self-referencing transaction or validation error)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: false}
                  message: {type: string, example: "Validation failed"}
                  data: {type: object, additionalProperties: {type: string}}
                  timestamp: {type: string, example: "2026-04-19T17:56:17"}
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: false}
                  message: {type: string, example: "An unexpected error occurred"}
                  timestamp: {type: string, example: "2026-04-19T17:56:17"}
  /transactions:
    get:
      tags:
      - Transactions
      summary: List transactions
      description: "Returns all transactions, optionally filtered by type"
      operationId: findAll
      parameters:
      - name: type
        in: query
        description: "Filter by transaction type (e.g. 'car', 'shopping')"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      description: Transaction response payload
                      properties:
                        id:
                          type: integer
                          format: int64
                          example: 1
                        amount:
                          type: number
                          format: double
                          example: 150.0
                        type:
                          type: string
                          example: car
                        parent_id:
                          type: integer
                          format: int64
                          example: 0
                  timestamp:
                    type: string
                    format: date-time
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: false}
                  message: {type: string, example: "An unexpected error occurred"}
                  timestamp: {type: string, example: "2026-04-19T17:56:17"}
  /transactions/types/{type}:
    get:
      tags:
      - Transactions
      summary: Get transaction ids by type
      description: Returns a list of ids of all transactions matching the given type
      operationId: findIdsByType
      parameters:
      - name: type
        in: path
        description: "Transaction type (e.g. 'car', 'shopping')"
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: integer
                      format: int64
                  timestamp:
                    type: string
                    format: date-time
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: false}
                  message: {type: string, example: "An unexpected error occurred"}
                  timestamp: {type: string, example: "2026-04-19T17:56:17"}
  /transactions/sum/{id}:
    get:
      tags:
      - Transactions
      summary: Get sum of a transaction subtree
      description: "Returns {\"sum\": <double>} with the total amount of the transaction\
        \ and all its transitive children"
      operationId: sum
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      sum:
                        type: number
                        format: double
                        example: 5000.0
                  timestamp:
                    type: string
                    format: date-time
                    example: "2026-04-19T19:38:38.447905335"
        "404":
          description: Transaction not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: false}
                  message: {type: string, example: "Transaction not found: 123"}
                  timestamp: {type: string, example: "2026-04-19T17:56:17"}
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: false}
                  message: {type: string, example: "An unexpected error occurred"}
                  timestamp: {type: string, example: "2026-04-19T17:56:17"}